+++ assets/application-admin.js
@@ -11,5 +11,5 @@ import './styles/application-admin.scss';
 import './../src/Blocks/assets/application-admin';
 
 // // Load Scripts
-import './scripts/application-admin';
+// import './scripts/application-admin';
 import './../src/Blocks/assets/scripts/application-admin';