+++ assets/application.js
@@ -6,14 +6,5 @@
  * Usage: `WordPress frontend screen`.
  */
 
-// Load Styles
-import './styles/application.scss';
-
-// // Load Scripts
-import './scripts/application';
-
 // Load images
 import './images';
-
-// Load fonts
-import './fonts';