+++ src/Blocks/components/form/assets/form.js
@@ -2,19 +2,22 @@
 
 import { cookies } from '@eightshift/frontend-libs/scripts/helpers';
 
+const ePrefix = 'esForms';
+
 // All custom events.
 export const FORM_EVENTS = {
-	BEFORE_FORM_SUBMIT: 'BeforeFormSubmit',
-	AFTER_FORM_SUBMIT: 'AfterFormSubmit',
-	AFTER_FORM_SUBMIT_SUCCESS_REDIRECT: 'AfterFormSubmitSuccessRedirect',
-	AFTER_FORM_SUBMIT_SUCCESS: 'AfterFormSubmitSuccess',
-	AFTER_FORM_SUBMIT_RESET: 'AfterFormSubmitReset',
-	AFTER_FORM_SUBMIT_ERROR: 'AfterFormSubmitError',
-	AFTER_FORM_SUBMIT_ERROR_FATAL: 'AfterFormSubmitErrorFatal',
-	AFTER_FORM_SUBMIT_ERROR_VALIDATION: 'AfterFormSubmitErrorValidation',
-	AFTER_FORM_SUBMIT_END: 'AfterFormSubmitEnd',
-	AFTER_FORM_EVENTS_CLEAR: 'AfterFormEventsClear',
-	BEFORE_GTM_DATA_PUSH: 'BeforeGtmDataPush',
+	BEFORE_FORM_SUBMIT: `${ePrefix}BeforeFormSubmit`,
+	AFTER_FORM_SUBMIT: `${ePrefix}AfterFormSubmit`,
+	AFTER_FORM_SUBMIT_SUCCESS_REDIRECT: `${ePrefix}AfterFormSubmitSuccessRedirect`,
+	AFTER_FORM_SUBMIT_SUCCESS: `${ePrefix}AfterFormSubmitSuccess`,
+	AFTER_FORM_SUBMIT_RESET: `${ePrefix}AfterFormSubmitReset`,
+	AFTER_FORM_SUBMIT_ERROR: `${ePrefix}AfterFormSubmitError`,
+	AFTER_FORM_SUBMIT_ERROR_FATAL: `${ePrefix}AfterFormSubmitErrorFatal`,
+	AFTER_FORM_SUBMIT_ERROR_VALIDATION: `${ePrefix}AfterFormSubmitErrorValidation`,
+	AFTER_FORM_SUBMIT_END: `${ePrefix}AfterFormSubmitEnd`,
+	AFTER_FORM_EVENTS_CLEAR: `${ePrefix}AfterFormEventsClear`,
+	BEFORE_GTM_DATA_PUSH: `${ePrefix}BeforeGtmDataPush`,
+	FORMS_JS_LOADED: `${ePrefix}JsLoaded`,
 };
 
 // All form custom state selectors.
@@ -153,7 +156,7 @@ export class Form {
 
 		// Set localStorage data from global variable.
 		this.setLocalStorage();
-	}
+	};
 
 	// Handle form submit and all logic.
 	onFormSubmit = (event) => {
@@ -170,7 +173,7 @@ export class Form {
 		} else {
 			this.formSubmit(element);
 		}
-	}
+	};
 
 	// Handle form submit and all logic for only one field click.
 	onFormSubmitSingle = (event) => {
@@ -178,7 +181,7 @@ export class Form {
 		const {target} = event;
 
 		this.formSubmit(target.closest(this.formSelector), target);
-	}
+	};
 
 	// Handle form submit and all logic in case we have captcha in place.
 	formSubmitCaptcha = (element, token) => {
@@ -230,7 +233,7 @@ export class Form {
 				}, parseInt(this.hideGlobalMessageTimeout, 10));
 			}
 		});
-	}
+	};
 
 	// Handle form submit and all logic.
 	formSubmit = (element, singleSubmit = false) => {
@@ -357,7 +360,7 @@ export class Form {
 				// Dispatch event.
 				this.dispatchFormEvent(element, FORM_EVENTS.AFTER_FORM_SUBMIT_END);
 			});
-	}
+	};
 
 	// Build form data object.
 	getFormData = (element, singleSubmit = false) => {
@@ -532,7 +535,7 @@ export class Form {
 	 }
 
 		return formData;
-	}
+	};
 
 	// Output all error for fields.
 	outputErrors = (element, fields) => {
@@ -553,7 +556,7 @@ export class Form {
 
 			this.scrollToElement(element.querySelector(`${this.errorSelector}[data-id="${firstItem}"]`).parentElement);
 		}
-	}
+	};
 
 	// Reset form values if the condition is right.
 	resetForm = (element) => {
@@ -590,7 +593,7 @@ export class Form {
 			// Dispatch event.
 			this.dispatchFormEvent(element, FORM_EVENTS.AFTER_FORM_SUBMIT_RESET);
 		}
-	}
+	};
 
 	// Reset for in general.
 	reset = (element) => {
@@ -603,7 +606,7 @@ export class Form {
 		element.querySelectorAll(`.${this.CLASS_HAS_ERROR}`).forEach((element) => element.classList.remove(this.CLASS_HAS_ERROR));
 
 		this.unsetGlobalMsg(element);
-	}
+	};
 
 	// Show loader.
 	showLoader = (element) => {
@@ -616,7 +619,7 @@ export class Form {
 		}
 
 		loader.classList.add(this.CLASS_ACTIVE);
-	}
+	};
 
 	// Hide loader.
 	hideLoader = (element) => {
@@ -631,7 +634,7 @@ export class Form {
 
 			loader.classList.remove(this.CLASS_ACTIVE);
 		}, parseInt(this.hideLoadingStateTimeout, 10));
-	}
+	};
 
 	// Set global message.
 	setGlobalMsg = (element, msg, status) => {
@@ -653,7 +656,7 @@ export class Form {
 		if (status === 'success' && this.formDisableScrollToGlobalMessageOnSuccess !== '1') {
 			this.scrollToElement(messageContainer);
 		}
-	}
+	};
 
 	// Unset global message.
 	unsetGlobalMsg(element) {
@@ -767,11 +770,11 @@ export class Form {
 		if (element !== null) {
 			element.scrollIntoView({block: 'start', behavior: 'smooth'});
 		}
-	}
+	};
 
 	// Dispatch custom event.
 	dispatchFormEvent(element, name) {
-		const event = new CustomEvent(`esForms${name}`, {
+		const event = new CustomEvent(name, {
 			bubbles: true
 		});
 
@@ -785,7 +788,7 @@ export class Form {
 		input.addEventListener('keydown', this.onFocusEvent);
 		input.addEventListener('focus', this.onFocusEvent);
 		input.addEventListener('blur', this.onBlurEvent);
-	}
+	};
 
 	// Setup Select field.
 	setupSelectField = (select, formId) => {
@@ -813,7 +816,7 @@ export class Form {
 			select.addEventListener('focus', this.onFocusEvent);
 			select.addEventListener('blur', this.onBlurEvent);
 		}
-	}
+	};
 
 	// Setup Textarea field.
 	setupTextareaField = (textarea, formId) => {
@@ -833,7 +836,7 @@ export class Form {
 				this.customTextareas[formId].push(autosize.default);
 			});
 		}
-	}
+	};
 
 	// Setup file single field.
 	setupFileField = (file, formId, index) => {
@@ -888,7 +891,7 @@ export class Form {
 				button.addEventListener('blur', this.onBlurEvent);
 			});
 		}
-	}
+	};
 
 	// On custom file wrapper click event callback.
 	onCustomFileWrapClickEvent = (event) => {
@@ -899,7 +902,7 @@ export class Form {
 		const formId = event.currentTarget.getAttribute('dropzone-form-id');
 
 		this.customFiles[formId][index].hiddenFileInput.click();
-	}
+	};
 
 	// Prefill inputs active/filled on init.
 	preFillOnInit = (input, type) => {
@@ -924,12 +927,12 @@ export class Form {
 				}
 				break;
 		}
-	}
+	};
 
 	// On Focus event for regular fields.
 	onFocusEvent = (event) => {
 		event.target.closest(this.fieldSelector).classList.add(this.CLASS_ACTIVE);
-	}
+	};
 
 	// On Blur generic method. Check for length of value.
 	onBlurEvent = (event) => {
@@ -972,7 +975,7 @@ export class Form {
 		} else {
 			field.classList.remove(this.CLASS_ACTIVE, this.CLASS_FILLED);
 		}
-	}
+	};
 
 	// Determine if field is custom type or normal.
 	isCustom(item) {
@@ -1070,7 +1073,18 @@ export class Form {
 		const searchParams = new URLSearchParams(window.location.search);
 
 		// Get storage from backend this is considered new by the page request.
-		const newStorage = searchParams.entries().filter(([key, value]) => allowedTags.includes(key) && value !== '');
+		const newStorage = {};
+
+		// Loop entries and get new storage values.
+		for (const [key, value] of searchParams.entries()) {
+			// Bailout if not allowed or empty
+			if (!allowedTags.includes(key) || value === '') {
+				continue;
+			}
+
+			// Add valid tag.
+			newStorage[key] = value;
+		}
 
 		// Bailout if nothing is set from allowed tags or everything is empty.
 		if (Object.keys(newStorage).length === 0) {