+++ src/Enqueue/Theme/EnqueueTheme.php
@@ -13,9 +13,7 @@ namespace EightshiftForms\Enqueue\Theme;
 use EightshiftForms\Config\Config;
 use EightshiftForms\Settings\Settings\SettingsGeneral;
 use EightshiftForms\Settings\SettingsHelper;
-use EightshiftForms\Hooks\Filters;
 use EightshiftForms\Hooks\Variables;
-use EightshiftForms\Tracking\TrackingInterface;
 use EightshiftForms\Validation\SettingsCaptcha;
 use EightshiftFormsVendor\EightshiftLibs\Manifest\ManifestInterface;
 use EightshiftFormsVendor\EightshiftLibs\Enqueue\Theme\AbstractEnqueueTheme;
@@ -31,22 +29,13 @@ class EnqueueTheme extends AbstractEnqueueTheme
 	use SettingsHelper;
 
 	/**
-	 * Instance variable of tracking data.
-	 *
-	 * @var TrackingInterface
-	 */
-	protected TrackingInterface $tracking;
-
-	/**
 	 * Create a new admin instance.
 	 *
 	 * @param ManifestInterface $manifest Inject manifest which holds data about assets from manifest.json.
-	 * @param TrackingInterface $tracking Inject tracking which holds data about for storing to localstorage.
 	 */
-	public function __construct(ManifestInterface $manifest, TrackingInterface $tracking)
+	public function __construct(ManifestInterface $manifest)
 	{
 		$this->manifest = $manifest;
-		$this->tracking = $tracking;
 	}
 
 	/**
@@ -56,8 +45,6 @@ class EnqueueTheme extends AbstractEnqueueTheme
 	 */
 	public function register(): void
 	{
-		\add_action('wp_enqueue_scripts', [$this, 'enqueueStylesLocal'], 10);
-		\add_action('wp_enqueue_scripts', [$this, 'enqueueScriptsLocal']);
 		\add_action('wp_enqueue_scripts', [$this, 'enqueueScriptsCaptcha']);
 	}
 
@@ -137,64 +124,5 @@ class EnqueueTheme extends AbstractEnqueueTheme
 	public function getAssetsVersion(): string
 	{
 		return Config::getProjectVersion();
-	}
-
-	/**
-	 * Get script localizations
-	 *
-	 * @return array<string, mixed>
-	 */
-	protected function getLocalizations(): array
-	{
-		$restRoutesPath = \rest_url() . Config::getProjectRoutesNamespace() . '/' . Config::getProjectRoutesVersion();
-
-		$hideGlobalMsgTimeoutFilterName = Filters::getBlockFilterName('form', 'hideGlobalMsgTimeout');
-		$redirectionTimeoutFilterName = Filters::getBlockFilterName('form', 'redirectionTimeout');
-		$previewRemoveLabelFilterName = Filters::getBlockFilterName('file', 'previewRemoveLabel');
-		$hideLoadingStateTimeoutFilterName = Filters::getBlockFilterName('form', 'hideLoadingStateTimeout');
-
-		$output = [
-			'formSubmitRestApiUrl' => $restRoutesPath . '/form-submit',
-			'hideGlobalMessageTimeout' => \apply_filters($hideGlobalMsgTimeoutFilterName, 6000),
-			'redirectionTimeout' => \apply_filters($redirectionTimeoutFilterName, 300),
-			'hideLoadingStateTimeout' => \apply_filters($hideLoadingStateTimeoutFilterName, 600),
-			'fileCustomRemoveLabel' => \apply_filters($previewRemoveLabelFilterName, \esc_html__('Remove', 'eightshift-forms')),
-			'formDisableScrollToFieldOnError' => $this->isCheckboxOptionChecked(
-				SettingsGeneral::SETTINGS_GENERAL_DISABLE_SCROLL_TO_FIELD_ON_ERROR,
-				SettingsGeneral::SETTINGS_GENERAL_DISABLE_SCROLL_KEY
-			),
-			'formDisableScrollToGlobalMessageOnSuccess' => $this->isCheckboxOptionChecked(
-				SettingsGeneral::SETTINGS_GENERAL_DISABLE_SCROLL_TO_GLOBAL_MESSAGE_ON_SUCCESS,
-				SettingsGeneral::SETTINGS_GENERAL_DISABLE_SCROLL_KEY
-			),
-			'formDisableAutoInit' => $this->isCheckboxOptionChecked(
-				SettingsGeneral::SETTINGS_GENERAL_DISABLE_AUTOINIT_ENQUEUE_SCRIPT_KEY,
-				SettingsGeneral::SETTINGS_GENERAL_DISABLE_DEFAULT_ENQUEUE_KEY
-			),
-			'formResetOnSuccess' => !Variables::isDevelopMode(),
-			'captcha' => '',
-			'storageConfig' => '',
-		];
-
-		// Check if Captcha data is set and valid.
-		$isCaptchaSettingsGlobalValid = \apply_filters(SettingsCaptcha::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME, false);
-
-		if ($isCaptchaSettingsGlobalValid) {
-			$output['captcha'] = !empty(Variables::getGoogleReCaptchaSiteKey()) ? Variables::getGoogleReCaptchaSiteKey() : $this->getOptionValue(SettingsCaptcha::SETTINGS_CAPTCHA_SITE_KEY);
-		}
-
-		// Localstorage allowed tags.
-		$allowedTrackingTags = $this->tracking->getAllowedTags();
-
-		if ($allowedTrackingTags) {
-			$output['storageConfig'] = \wp_json_encode([
-				'allowed' => $allowedTrackingTags,
-				'expiration' => $this->tracking->getTrackingExpiration(),
-			]);
-		}
-
-		return [
-			'esFormsLocalization' => $output,
-		];
 	}
 }